+++ objects/src/block/mod.rs
@@ -102,16 +102,20 @@ impl Block {
     pub fn notes(&self) -> impl Iterator<Item = (BlockNoteIndex, &OutputNote)> {
         self.output_note_batches.iter().enumerate().flat_map(|(batch_idx, notes)| {
             notes.iter().enumerate().map(move |(note_idx_in_batch, note)| {
-                (BlockNoteIndex::new(batch_idx, note_idx_in_batch), note)
+                (
+                    BlockNoteIndex::new(batch_idx, note_idx_in_batch).expect(
+                        "Something went wrong: block is invalid, but passed or skipped validation",
+                    ),
+                    note,
+                )
             })
         })
     }
 
     /// Returns a note tree containing all notes created in this block.
     pub fn build_note_tree(&self) -> BlockNoteTree {
-        let entries = self
-            .notes()
-            .map(|(note_index, note)| (note_index, note.id().into(), *note.metadata()));
+        let entries =
+            self.notes().map(|(note_index, note)| (note_index, note.id(), *note.metadata()));
 
         BlockNoteTree::with_entries(entries)
             .expect("Something went wrong: block is invalid, but passed or skipped validation")