+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - [BREAKING] `AccountStub` structure was renamed to `AccountHeader` (#855).
 - [BREAKING] Kernel procedures now have to be invoked using `dynexec` instruction (#803).
 - Refactored `AccountStorage` from `Smt` to `sequential hash` (#846)
+- [BREAKING] Refactored batch/block note trees (#834).
 
 ## 0.5.1 (2024-08-28) - `miden-objects` crate only